Add tutorial notebook auto test/export.
Running the tutorials.py file as from the utils directory will
run all the tutorials in the configuration file. More tutorials
will be added later to the configuration file. The tutorials will
run with all the different configurations in the configuration file
and export the output using the custom RST template. The output
directory is the tutorials directory (chipwhisperer-tutorial submodule).