Dockerfile, startup, and mail server now stable.
The docker build works now. It includes a send-only mail server
for sending the results of the tests to your e-mail. Pretty much
guaranteed to land in spam, so check there. The tracebacks are still
not transferred over to plain text well (they are gibberish). That
can be fixed by changing the tutorials.py.