PA_SPA_1: Fix checkpass function
Seems like there is a mixup when it comes to less than and greater than
in the "Attacking a Single Letter" example. For example for
"CW303" and "CWLITEXMEGA" there is a comparison like this

   return trace[73 + 40 * i] > -0.3

Since we are interested in (negative) peaks in that case we should use a
less than sign instead of greater than.

After doing the changes in this patch I've been able to successfully run
the example in Jupyter using Chipwhisperer-Lite ("CW303", "CWLITEXMEGA").

Since I believe it's wrong for "CWNANO" as well as for "CWLITEARM" and
"CW308_STM32F3" I've made similar changes.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>